Refactor: extract proof gen and leaf functions to new files; Add more unit tests.
