+++ merkle_tree_test.go
@@ -26,7 +26,6 @@ import (
 	"bytes"
 	"crypto/rand"
 	"crypto/sha256"
-	"encoding/hex"
 	"errors"
 	"fmt"
 	"reflect"
@@ -39,7 +38,7 @@ import (
 
 const benchSize = 65536
 
-func generatedTestDataBlocks(num int) []DataBlock {
+func mockDataBlocks(num int) []DataBlock {
 	blocks := make([]DataBlock, num)
 	for i := 0; i < num; i++ {
 		block := &mock.DataBlock{
@@ -51,255 +50,6 @@ func generatedTestDataBlocks(num int) []DataBlock {
 		blocks[i] = block
 	}
 	return blocks
-}
-
-func TestMerkleTreeNew_modeProofGen(t *testing.T) {
-	dummyDataBlocks := []DataBlock{
-		&mock.DataBlock{
-			Data: []byte("dummy_data_0"),
-		},
-		&mock.DataBlock{
-			Data: []byte("dummy_data_1"),
-		},
-		&mock.DataBlock{
-			Data: []byte("dummy_data_2"),
-		},
-		&mock.DataBlock{
-			Data: []byte("dummy_data_3"),
-		},
-		&mock.DataBlock{
-			Data: []byte("dummy_data_4"),
-		},
-	}
-	dummyRootSizeTwo, err := hex.DecodeString("30c87249cdfa43ed48a8e6a7747e05312933eaf2cbca38974e101649d3bae339")
-	if err != nil {
-		t.Fatal(err)
-	}
-	dummyRootSizeThree, err := hex.DecodeString("1437da40a42b98b5fd7a76493375a687f0b2418071a009d1459f24e61db48c70")
-	if err != nil {
-		t.Fatal(err)
-	}
-	dummyRootSizeFour, err := hex.DecodeString("bfef6f8177eaf488386177d31a24e0481781811388ac1c8e291d564dc3a79fb9")
-	if err != nil {
-		t.Fatal(err)
-	}
-	dummyRootSizeFive, err := hex.DecodeString("5e9ba12514859d30d6e79a6f2d6e9aa71a6b011ae35e6c07f487c6d570993d71")
-	if err != nil {
-		t.Fatal(err)
-	}
-	type args struct {
-		blocks []DataBlock
-		config *Config
-	}
-	tests := []struct {
-		name     string
-		args     args
-		wantErr  bool
-		wantRoot []byte
-	}{
-		{
-			name: "test_0",
-			args: args{
-				blocks: generatedTestDataBlocks(0),
-			},
-			wantErr: true,
-		},
-		{
-			name: "test_1",
-			args: args{
-				blocks: generatedTestDataBlocks(1),
-			},
-			wantErr: true,
-		},
-		{
-			name: "test_2",
-			args: args{
-				blocks: []DataBlock{dummyDataBlocks[0], dummyDataBlocks[1]},
-			},
-			wantErr:  false,
-			wantRoot: dummyRootSizeTwo,
-		},
-		{
-			name: "test_3",
-			args: args{
-				blocks: []DataBlock{dummyDataBlocks[0], dummyDataBlocks[1], dummyDataBlocks[2]},
-			},
-			wantErr:  false,
-			wantRoot: dummyRootSizeThree,
-		},
-		{
-			name: "test_4",
-			args: args{
-				blocks: []DataBlock{dummyDataBlocks[0], dummyDataBlocks[1], dummyDataBlocks[2], dummyDataBlocks[3]},
-			},
-			wantErr:  false,
-			wantRoot: dummyRootSizeFour,
-		},
-		{
-			name: "test_5",
-			args: args{
-				blocks: dummyDataBlocks,
-			},
-			wantErr:  false,
-			wantRoot: dummyRootSizeFive,
-		},
-		{
-			name: "test_7",
-			args: args{
-				blocks: generatedTestDataBlocks(7),
-			},
-			wantErr: false,
-		},
-		{
-			name: "test_8",
-			args: args{
-				blocks: generatedTestDataBlocks(8),
-			},
-			wantErr: false,
-		},
-		{
-			name: "test_5",
-			args: args{
-				blocks: generatedTestDataBlocks(5),
-			},
-			wantErr: false,
-		},
-		{
-			name: "test_1000",
-			args: args{
-				blocks: generatedTestDataBlocks(1000),
-			},
-			wantErr: false,
-		},
-		{
-			name: "test_100_parallel_4",
-			args: args{
-				blocks: generatedTestDataBlocks(100),
-				config: &Config{
-					RunInParallel: true,
-					NumRoutines:   4,
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "test_10_parallel_32",
-			args: args{
-				blocks: generatedTestDataBlocks(100),
-				config: &Config{
-					RunInParallel: true,
-					NumRoutines:   32,
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "test_100_parallel_no_specify_num_of_routines",
-			args: args{
-				blocks: generatedTestDataBlocks(100),
-				config: &Config{
-					RunInParallel: true,
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "test_8_sorted",
-			args: args{
-				blocks: generatedTestDataBlocks(8),
-				config: &Config{
-					SortSiblingPairs: true,
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "test_hash_func_error",
-			args: args{
-				blocks: generatedTestDataBlocks(100),
-				config: &Config{
-					HashFunc: func([]byte) ([]byte, error) {
-						return nil, fmt.Errorf("hash func error")
-					},
-				},
-			},
-			wantErr: true,
-		},
-		{
-			name: "test_hash_func_error_parallel",
-			args: args{
-				blocks: generatedTestDataBlocks(100),
-				config: &Config{
-					HashFunc: func([]byte) ([]byte, error) {
-						return nil, fmt.Errorf("hash func error")
-					},
-					RunInParallel: true,
-				},
-			},
-			wantErr: true,
-		},
-		{
-			name: "test_100_disable_leaf_hashing",
-			args: args{
-				blocks: generatedTestDataBlocks(100),
-				config: &Config{
-					DisableLeafHashing: true,
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "test_100_disable_leaf_hashing_parallel_4",
-			args: args{
-				blocks: generatedTestDataBlocks(100),
-				config: &Config{
-					DisableLeafHashing: true,
-					RunInParallel:      true,
-					NumRoutines:        4,
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "invalid_mode",
-			args: args{
-				blocks: generatedTestDataBlocks(100),
-				config: &Config{
-					Mode: 5,
-				},
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			mt, err := New(tt.args.config, tt.args.blocks)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("Build() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if tt.wantErr {
-				return
-			}
-			if mt == nil {
-				t.Errorf("Build() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if tt.wantRoot == nil {
-				for idx, block := range tt.args.blocks {
-					if ok, _ := mt.Verify(block, mt.Proofs[idx]); !ok {
-						t.Errorf("proof verification failed, idx %d", idx)
-						return
-					}
-				}
-			} else {
-				if !bytes.Equal(mt.Root, tt.wantRoot) {
-					t.Errorf("root mismatch, got %x, want %x", mt.Root, tt.wantRoot)
-					return
-				}
-			}
-		})
-	}
 }
 
 func TestMerkleTreeNew_modeTreeBuild(t *testing.T) {
@@ -316,7 +66,7 @@ func TestMerkleTreeNew_modeTreeBuild(t *testing.T) {
 		{
 			name: "test_build_tree_2",
 			args: args{
-				blocks: generatedTestDataBlocks(2),
+				blocks: mockDataBlocks(2),
 				config: &Config{
 					Mode: ModeTreeBuild,
 				},
@@ -326,7 +76,7 @@ func TestMerkleTreeNew_modeTreeBuild(t *testing.T) {
 		{
 			name: "test_build_tree_3",
 			args: args{
-				blocks: generatedTestDataBlocks(3),
+				blocks: mockDataBlocks(3),
 				config: &Config{
 					Mode: ModeTreeBuild,
 				},
@@ -336,7 +86,7 @@ func TestMerkleTreeNew_modeTreeBuild(t *testing.T) {
 		{
 			name: "test_build_tree_5",
 			args: args{
-				blocks: generatedTestDataBlocks(5),
+				blocks: mockDataBlocks(5),
 				config: &Config{
 					Mode: ModeTreeBuild,
 				},
@@ -346,7 +96,7 @@ func TestMerkleTreeNew_modeTreeBuild(t *testing.T) {
 		{
 			name: "test_build_tree_8",
 			args: args{
-				blocks: generatedTestDataBlocks(8),
+				blocks: mockDataBlocks(8),
 				config: &Config{
 					Mode: ModeTreeBuild,
 				},
@@ -356,7 +106,7 @@ func TestMerkleTreeNew_modeTreeBuild(t *testing.T) {
 		{
 			name: "test_build_tree_16",
 			args: args{
-				blocks: generatedTestDataBlocks(16),
+				blocks: mockDataBlocks(16),
 				config: &Config{
 					Mode: ModeTreeBuild,
 				},
@@ -366,7 +116,7 @@ func TestMerkleTreeNew_modeTreeBuild(t *testing.T) {
 		{
 			name: "test_build_tree_32",
 			args: args{
-				blocks: generatedTestDataBlocks(32),
+				blocks: mockDataBlocks(32),
 				config: &Config{
 					Mode: ModeTreeBuild,
 				},
@@ -376,7 +126,7 @@ func TestMerkleTreeNew_modeTreeBuild(t *testing.T) {
 		{
 			name: "test_build_tree_36",
 			args: args{
-				blocks: generatedTestDataBlocks(36),
+				blocks: mockDataBlocks(36),
 				config: &Config{
 					Mode: ModeTreeBuild,
 				},
@@ -386,7 +136,7 @@ func TestMerkleTreeNew_modeTreeBuild(t *testing.T) {
 		{
 			name: "test_build_tree_1000",
 			args: args{
-				blocks: generatedTestDataBlocks(1000),
+				blocks: mockDataBlocks(1000),
 				config: &Config{
 					Mode: ModeTreeBuild,
 				},
@@ -396,7 +146,7 @@ func TestMerkleTreeNew_modeTreeBuild(t *testing.T) {
 		{
 			name: "test_hash_func_error",
 			args: args{
-				blocks: generatedTestDataBlocks(100),
+				blocks: mockDataBlocks(100),
 				config: &Config{
 					HashFunc: func([]byte) ([]byte, error) {
 						return nil, fmt.Errorf("hash func error")
@@ -409,7 +159,7 @@ func TestMerkleTreeNew_modeTreeBuild(t *testing.T) {
 		{
 			name: "test_disable_leaf_hashing",
 			args: args{
-				blocks: generatedTestDataBlocks(100),
+				blocks: mockDataBlocks(100),
 				config: &Config{
 					DisableLeafHashing: true,
 					Mode:               ModeTreeBuild,
@@ -456,7 +206,7 @@ func TestMerkleTreeNew_modeTreeBuildRunInParallel(t *testing.T) {
 		{
 			name: "test_build_tree_parallel_2",
 			args: args{
-				blocks: generatedTestDataBlocks(2),
+				blocks: mockDataBlocks(2),
 				config: &Config{
 					RunInParallel: true,
 					NumRoutines:   4,
@@ -468,7 +218,7 @@ func TestMerkleTreeNew_modeTreeBuildRunInParallel(t *testing.T) {
 		{
 			name: "test_build_tree_parallel_4",
 			args: args{
-				blocks: generatedTestDataBlocks(4),
+				blocks: mockDataBlocks(4),
 				config: &Config{
 					RunInParallel: true,
 					NumRoutines:   4,
@@ -480,7 +230,7 @@ func TestMerkleTreeNew_modeTreeBuildRunInParallel(t *testing.T) {
 		{
 			name: "test_build_tree_parallel_5",
 			args: args{
-				blocks: generatedTestDataBlocks(5),
+				blocks: mockDataBlocks(5),
 				config: &Config{
 					RunInParallel: true,
 					NumRoutines:   4,
@@ -492,7 +242,7 @@ func TestMerkleTreeNew_modeTreeBuildRunInParallel(t *testing.T) {
 		{
 			name: "test_build_tree_parallel_8",
 			args: args{
-				blocks: generatedTestDataBlocks(8),
+				blocks: mockDataBlocks(8),
 				config: &Config{
 					RunInParallel: true,
 					NumRoutines:   4,
@@ -504,7 +254,7 @@ func TestMerkleTreeNew_modeTreeBuildRunInParallel(t *testing.T) {
 		{
 			name: "test_build_tree_parallel_8_32",
 			args: args{
-				blocks: generatedTestDataBlocks(8),
+				blocks: mockDataBlocks(8),
 				config: &Config{
 					RunInParallel: true,
 					NumRoutines:   32,
@@ -516,7 +266,7 @@ func TestMerkleTreeNew_modeTreeBuildRunInParallel(t *testing.T) {
 		{
 			name: "test_hash_func_error_parallel",
 			args: args{
-				blocks: generatedTestDataBlocks(100),
+				blocks: mockDataBlocks(100),
 				config: &Config{
 					HashFunc: func([]byte) ([]byte, error) {
 						return nil, fmt.Errorf("hash func error")
@@ -563,7 +313,7 @@ func TestMerkleTreeNew_modeProofGenAndTreeBuild(t *testing.T) {
 		{
 			name: "test_build_tree_proof_2",
 			args: args{
-				blocks: generatedTestDataBlocks(2),
+				blocks: mockDataBlocks(2),
 				config: &Config{
 					Mode: ModeProofGenAndTreeBuild,
 				},
@@ -573,7 +323,7 @@ func TestMerkleTreeNew_modeProofGenAndTreeBuild(t *testing.T) {
 		{
 			name: "test_build_tree_proof_4",
 			args: args{
-				blocks: generatedTestDataBlocks(4),
+				blocks: mockDataBlocks(4),
 				config: &Config{
 					Mode: ModeProofGenAndTreeBuild,
 				},
@@ -583,7 +333,7 @@ func TestMerkleTreeNew_modeProofGenAndTreeBuild(t *testing.T) {
 		{
 			name: "test_build_tree_proof_5",
 			args: args{
-				blocks: generatedTestDataBlocks(5),
+				blocks: mockDataBlocks(5),
 				config: &Config{
 					Mode: ModeProofGenAndTreeBuild,
 				},
@@ -593,7 +343,7 @@ func TestMerkleTreeNew_modeProofGenAndTreeBuild(t *testing.T) {
 		{
 			name: "test_build_tree_proof_8",
 			args: args{
-				blocks: generatedTestDataBlocks(8),
+				blocks: mockDataBlocks(8),
 				config: &Config{
 					Mode: ModeProofGenAndTreeBuild,
 				},
@@ -603,7 +353,7 @@ func TestMerkleTreeNew_modeProofGenAndTreeBuild(t *testing.T) {
 		{
 			name: "test_build_tree_proof_9",
 			args: args{
-				blocks: generatedTestDataBlocks(9),
+				blocks: mockDataBlocks(9),
 				config: &Config{
 					Mode: ModeProofGenAndTreeBuild,
 				},
@@ -613,7 +363,7 @@ func TestMerkleTreeNew_modeProofGenAndTreeBuild(t *testing.T) {
 		{
 			name: "test_hash_func_error",
 			args: args{
-				blocks: generatedTestDataBlocks(100),
+				blocks: mockDataBlocks(100),
 				config: &Config{
 					HashFunc: func([]byte) ([]byte, error) {
 						return nil, fmt.Errorf("hash func error")
@@ -626,7 +376,7 @@ func TestMerkleTreeNew_modeProofGenAndTreeBuild(t *testing.T) {
 		{
 			name: "test_tree_build_hash_func_error",
 			args: args{
-				blocks: generatedTestDataBlocks(100),
+				blocks: mockDataBlocks(100),
 				config: &Config{
 					HashFunc: func(block []byte) ([]byte, error) {
 						if len(block) == 64 {
@@ -680,7 +430,7 @@ func TestMerkleTreeNew_modeProofGenAndTreeBuildRunInParallel(t *testing.T) {
 		{
 			name: "test_build_tree_proof_parallel_2",
 			args: args{
-				blocks: generatedTestDataBlocks(2),
+				blocks: mockDataBlocks(2),
 				config: &Config{
 					RunInParallel: true,
 					NumRoutines:   4,
@@ -692,7 +442,7 @@ func TestMerkleTreeNew_modeProofGenAndTreeBuildRunInParallel(t *testing.T) {
 		{
 			name: "test_build_tree_proof_parallel_4",
 			args: args{
-				blocks: generatedTestDataBlocks(4),
+				blocks: mockDataBlocks(4),
 				config: &Config{
 					RunInParallel: true,
 					NumRoutines:   4,
@@ -704,7 +454,7 @@ func TestMerkleTreeNew_modeProofGenAndTreeBuildRunInParallel(t *testing.T) {
 		{
 			name: "test_build_tree_proof_parallel_5",
 			args: args{
-				blocks: generatedTestDataBlocks(5),
+				blocks: mockDataBlocks(5),
 				config: &Config{
 					RunInParallel: true,
 					NumRoutines:   4,
@@ -716,7 +466,7 @@ func TestMerkleTreeNew_modeProofGenAndTreeBuildRunInParallel(t *testing.T) {
 		{
 			name: "test_build_tree_proof_parallel_8",
 			args: args{
-				blocks: generatedTestDataBlocks(8),
+				blocks: mockDataBlocks(8),
 				config: &Config{
 					RunInParallel: true,
 					NumRoutines:   4,
@@ -728,7 +478,7 @@ func TestMerkleTreeNew_modeProofGenAndTreeBuildRunInParallel(t *testing.T) {
 		{
 			name: "test_hash_func_error",
 			args: args{
-				blocks: generatedTestDataBlocks(100),
+				blocks: mockDataBlocks(100),
 				config: &Config{
 					HashFunc: func([]byte) ([]byte, error) {
 						return nil, fmt.Errorf("hash func error")
@@ -742,7 +492,7 @@ func TestMerkleTreeNew_modeProofGenAndTreeBuildRunInParallel(t *testing.T) {
 		{
 			name: "test_tree_build_hash_func_error",
 			args: args{
-				blocks: generatedTestDataBlocks(100),
+				blocks: mockDataBlocks(100),
 				config: &Config{
 					HashFunc: func(block []byte) ([]byte, error) {
 						if len(block) == 64 {
@@ -785,7 +535,7 @@ func TestMerkleTreeNew_modeProofGenAndTreeBuildRunInParallel(t *testing.T) {
 }
 
 func setupTestVerify(size int) (*MerkleTree, []DataBlock) {
-	blocks := generatedTestDataBlocks(size)
+	blocks := mockDataBlocks(size)
 	m, err := New(nil, blocks)
 	if err != nil {
 		panic(err)
@@ -794,7 +544,7 @@ func setupTestVerify(size int) (*MerkleTree, []DataBlock) {
 }
 
 func setupTestVerifyRunInParallel(size int) (*MerkleTree, []DataBlock) {
-	blocks := generatedTestDataBlocks(size)
+	blocks := mockDataBlocks(size)
 	m, err := New(&Config{
 		RunInParallel: true,
 		NumRoutines:   1,
@@ -929,28 +679,28 @@ func TestMerkleTree_Proof(t *testing.T) {
 		{
 			name:   "test_2",
 			config: &Config{Mode: ModeTreeBuild},
-			blocks: generatedTestDataBlocks(2),
+			blocks: mockDataBlocks(2),
 		},
 		{
 			name:   "test_4",
 			config: &Config{Mode: ModeTreeBuild},
-			blocks: generatedTestDataBlocks(4),
+			blocks: mockDataBlocks(4),
 		},
 		{
 			name:   "test_5",
 			config: &Config{Mode: ModeTreeBuild},
-			blocks: generatedTestDataBlocks(5),
+			blocks: mockDataBlocks(5),
 		},
 		{
 			name:    "test_wrong_mode",
 			config:  &Config{Mode: ModeProofGen},
-			blocks:  generatedTestDataBlocks(5),
+			blocks:  mockDataBlocks(5),
 			wantErr: true,
 		},
 		{
 			name:   "test_wrong_blocks",
 			config: &Config{Mode: ModeTreeBuild},
-			blocks: generatedTestDataBlocks(5),
+			blocks: mockDataBlocks(5),
 			proofBlocks: []DataBlock{
 				&mock.DataBlock{
 					Data: []byte("test_wrong_blocks"),
@@ -967,7 +717,7 @@ func TestMerkleTree_Proof(t *testing.T) {
 						return nil, errors.New("data block serialize error")
 					})
 			},
-			blocks:  generatedTestDataBlocks(5),
+			blocks:  mockDataBlocks(5),
 			wantErr: true,
 		},
 	}
@@ -1008,62 +758,8 @@ func TestMerkleTree_Proof(t *testing.T) {
 	}
 }
 
-func mockHashFunc(data []byte) ([]byte, error) {
-	sha256Func := sha256.New()
-	sha256Func.Write(data)
-	return sha256Func.Sum(nil), nil
-}
-
-func TestMerkleTree_generateProofs(t *testing.T) {
-	patches := gomonkey.NewPatches()
-	defer patches.Reset()
-	type args struct {
-		config *Config
-		blocks []DataBlock
-	}
-	tests := []struct {
-		name    string
-		args    args
-		mock    func()
-		wantErr bool
-	}{
-		{
-			name: "test_hash_func_err",
-			args: args{
-				config: &Config{
-					HashFunc: mockHashFunc,
-				},
-				blocks: generatedTestDataBlocks(5),
-			},
-			mock: func() {
-				patches.ApplyFunc(mockHashFunc,
-					func([]byte) ([]byte, error) {
-						return nil, errors.New("test_hash_func_err")
-					})
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			m, err := New(tt.args.config, tt.args.blocks)
-			if err != nil {
-				t.Errorf("New() error = %v", err)
-				return
-			}
-			if tt.mock != nil {
-				tt.mock()
-			}
-			defer patches.Reset()
-			if err := m.generateProofs(); (err != nil) != tt.wantErr {
-				t.Errorf("generateProofs() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
 func TestVerify(t *testing.T) {
-	blocks := generatedTestDataBlocks(5)
+	blocks := mockDataBlocks(5)
 	m, err := New(nil, blocks)
 	if err != nil {
 		t.Errorf("New() error = %v", err)
@@ -1222,7 +918,7 @@ func TestVerify(t *testing.T) {
 }
 
 func BenchmarkMerkleTreeNew(b *testing.B) {
-	testCases := generatedTestDataBlocks(benchSize)
+	testCases := mockDataBlocks(benchSize)
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		_, err := New(nil, testCases)
@@ -1236,7 +932,7 @@ func BenchmarkMerkleTreeNew_modeRunInParallel(b *testing.B) {
 	config := &Config{
 		RunInParallel: true,
 	}
-	testCases := generatedTestDataBlocks(benchSize)
+	testCases := mockDataBlocks(benchSize)
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		_, err := New(config, testCases)
@@ -1247,7 +943,7 @@ func BenchmarkMerkleTreeNew_modeRunInParallel(b *testing.B) {
 }
 
 func BenchmarkMerkleTreeNew_modeTreeBuild(b *testing.B) {
-	testCases := generatedTestDataBlocks(benchSize)
+	testCases := mockDataBlocks(benchSize)
 	config := &Config{
 		Mode: ModeTreeBuild,
 	}
@@ -1265,7 +961,7 @@ func BenchmarkMerkleTreeNew_modeTreeBuildRunInParallel(b *testing.B) {
 		Mode:          ModeTreeBuild,
 		RunInParallel: true,
 	}
-	testCases := generatedTestDataBlocks(benchSize)
+	testCases := mockDataBlocks(benchSize)
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		_, err := New(config, testCases)
